pyproject.toml
[build-system]
requires = ["setuptools>=65"]
build-backend = "setuptools.build_meta"

[project]
name = "ns8lab"
version = "0.0.1"
description = "NS8 supervised learning lab"
authors = [{ name = "Your Name" }]
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
  "numpy==2.2.6",
  "pandas",
  "scikit-learn==1.8.0",
  "matplotlib",
  "joblib",
  "pyyaml"
]

[project.optional-dependencies]
dev = [
  "pytest>=7",
]

[tool.setuptools]
package-dir = {"" = "src"}

[tool.setuptools.packages.find]
where = ["src"]

[tool.pytest.ini_options]
pythonpath = ["src"]
addopts = "-q"
